add: anchor sync

+++ firmware/utils.h
@@ -1,9 +1,11 @@
 #ifndef UTILS_H
 #define UTILS_H
 
+#include <stdint.h>
+
 void utils_init();
 
 void utils_start_execution_timer();
-double utils_stop_execution_timer();
+uint32_t utils_stop_execution_timer();
 
 #endif // UTILS_H
